+++ src/i18ndude/untranslated.py
@@ -137,6 +137,12 @@ def attr_validator(tag, attrs, logfct):
             logfct('value attribute of <... submit/button> lacks '
                    'i18n:attribute', 'ERROR')
 
+    # 4)
+    if tag == 'input' and 'placeholder' in attrs.keys():
+        if not _valid_i18ned_attr('placeholder', attrs):
+            logfct('placeholder attribute of <%s> lacks i18n:attributes' % tag,
+                   'ERROR')
+
 
 class Handler(xml.sax.ContentHandler):